+++ src/seamlesserverpkg/db/setup.nim
@@ -1,0 +1,15 @@
+import ../db
+import ../db/models/user
+
+proc setup*(conn: DbConn) =
+  ## Creates all tables
+  conn.createTables newUser()
+
+  block createAdmin:
+    var user = newUser(
+      username = "admin",
+      email = "admin@localhost",
+      password = "admin",
+      kind = ukAdmin
+    )
+    conn.insert user